+++ src/scorer/feature/interactions_scorer.ts
@@ -6,7 +6,7 @@ import { mastodon } from "masto";
 import Account from "../../api/objects/account";
 import AccountScorer from "../acccount_scorer";
 import { countValues } from "../../helpers/collection_helpers";
-import { MastoApi } from "../../api/api";
+import MastoApi from "../../api/api";
 import { StringNumberDict, WeightName } from "../../types";