+++ src/scorer/feature/mentions_followed_scorer.ts
@@ -5,7 +5,7 @@ import Account from '../../api/objects/account';
 import FeatureScorer from '../feature_scorer';
 import Toot from '../../api/objects/toot';
 import { countValues } from '../../helpers/collection_helpers';
-import { MastoApi } from '../../api/api';
+import MastoApi from '../../api/api';
 import { StringNumberDict, WeightName } from '../../types';