+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Lower Config.maxRecordsForFeatureScoring to 1600, numTrendingTagsToots to 200
 * `HashtagParticipationScorer` uses the square root of the number of participations to tamp down runaway scores.
 * Use a Semaphore to limit concurrency of hashtag search API requests; add `Config.maxConcurrentTootRequests`
 * Exclude followed tags from the pulling of hashtag participation toots (followed tags are already in timeline)